refactor: Move Solo12Config to separate library
This makes it easier to use it in other libraries.

Also add Eigen to `find_package` and link libraries (also works without
but seems more correct to add it).
